refactor(tests): move detect_sources GitLab tests into tests/nodes/plan_actions/
Merges app/nodes/plan_actions/detect_sources_test.py into
tests/nodes/plan_actions/test_detect_sources.py and removes
the inline copy. All merge_request_iid assertions are preserved
unchanged. Closes #902.
